+++ app/renderer/actions/ServerMonitor.js
@@ -1,5 +1,6 @@
 import { ipcRenderer, shell } from 'electron';
 import { push } from 'react-router-redux';
+import { fs } from 'appium-support';
 
 export const SERVER_STOP_REQ = 'SERVER_STOP_REQ';
 export const SERVER_STOP_OK = 'SERVER_STOP_OK';
@@ -77,7 +78,11 @@ export function closeMonitor () {
 }
 
 export function clearLogs () {
-  return (dispatch) => {
+  return async (dispatch, getState) => {
+    const logfilePath = getState().startServer.logfilePath;
+    if (logfilePath) {
+      await fs.writeFile(logfilePath, '');
+    }
     dispatch({type: LOGS_CLEARED});
   };
 }