+++ app/renderer/components/ServerMonitor/ServerMonitor.js
@@ -105,6 +105,11 @@ class GetRawLogsButton extends Component {
 
 export default class ServerMonitor extends Component {
 
+  constructor (props) {
+    super(props);
+    this.keydownListener = this.keydownListener.bind(this);
+  }
+
   static propTypes = {
     stopServer: PropTypes.func.isRequired,
     closeMonitor: PropTypes.func.isRequired,
@@ -113,6 +118,20 @@ export default class ServerMonitor extends Component {
     startSession: PropTypes.func.isRequired,
     startPlayback: PropTypes.func.isRequired,
     serverArgs: PropTypes.object.isRequired,
+  }
+
+  keydownListener (e) {
+    if (e.key === 'k' && (e.metaKey || e.ctrlKey)) {
+      this.props.clearLogs();
+    }
+  }
+
+  componentDidMount () {
+    document.addEventListener('keydown', this.keydownListener);
+  }
+
+  componentWillUnmount () {
+    document.removeEventListener('keydown', this.keydownListener);
   }
 
   componentWillUpdate () {