+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+## [1.4.1](https://github.com/appium/appium-desktop/compare/v1.4.0...v1.4.1) (2018-03-06)
+
+### Changes
+* Fix: Replace electron-builder autoupdater with Zeit Hazel server plus native updater (#435)
+
+
+## [1.4.0](https://github.com/appium/appium-desktop/compare/v1.3.2...v1.4.0) (2018-02-27)
+
+### Changes
+* Feature: Add Headspin cloud to list of remote server types
+* Feature: Add keyboard shortcut to clear terminal logs (#419)
+* Fix: Fix Python recording error (#422)
+
+## [1.3.2](https://github.com/appium/appium-desktop/compare/v1.3.1...v1.3.2) (2018-01-30)
+
+### Changes
+* Fix: Revert to Electron 1.7.11
+
 ## [1.3.1](https://github.com/appium/appium-desktop/compare/v1.3.0...v1.3.1) (2017-01-11)
 
 ### Changes